Revert IPreferenceStore rename
Rename IPreferenceStore to PreferenceStore, to match the other
interfaces in the repo
