+++ app/src/main/java/org/jellyfin/androidtv/preference/BasicPreferenceStore.kt
@@ -1,37 +1,6 @@
 package org.jellyfin.androidtv.preference
 
-abstract class BasicPreferenceStore : IPreferenceStore {
-
-	override fun <T : Any> getDefaultValue(preference: Preference<T>): T {
-		return preference.defaultValue
-	}
-
-	override fun <T : Any> reset(preference: Preference<T>) {
-		this[preference] = getDefaultValue(preference)
-	}
-
-	@Suppress("UNCHECKED_CAST")
-	override operator fun <T : Any> get(preference: Preference<T>): T =
-		// Coerce returned type based on the default type
-		when (preference.defaultValue) {
-			is Int -> getInt(preference.key, preference.defaultValue)
-			is Long -> getLong(preference.key, preference.defaultValue)
-			is Boolean -> getBool(preference.key, preference.defaultValue)
-			is String -> getString(preference.key, preference.defaultValue)
-			else -> throw IllegalArgumentException("${preference.type.simpleName} type is not supported")
-		} as T
-
-	override operator fun <T : Enum<T>> get(preference: Preference<T>) = getEnum(preference)
-
-	override operator fun <T : Any> set(preference: Preference<T>, value: T) =
-		when (value) {
-			is Int -> setInt(preference.key, value)
-			is Long -> setLong(preference.key, value)
-			is Boolean -> setBool(preference.key, value)
-			is String -> setString(preference.key, value)
-			is Enum<*> -> setEnum(preference, value)
-			else -> throw IllegalArgumentException("${preference.type.simpleName} type is not supported")
-		}
+abstract class BasicPreferenceStore : PreferenceStore {
 
 	// Protected methods to get / set items, this is an implementation detail so we protect
 	// it in the abstract common functionality (where it is used)
