+++ app/src/main/java/org/jellyfin/androidtv/ui/browsing/DisplayPreferencesScreen.kt
@@ -4,7 +4,7 @@ import org.jellyfin.androidtv.R
 import org.jellyfin.androidtv.constant.GridDirection
 import org.jellyfin.androidtv.constant.ImageType
 import org.jellyfin.androidtv.constant.PosterSize
-import org.jellyfin.androidtv.preference.IPreferenceStore
+import org.jellyfin.androidtv.preference.PreferenceStore
 import org.jellyfin.androidtv.preference.LibraryPreferences
 import org.jellyfin.androidtv.preference.PreferencesRepository
 import org.jellyfin.androidtv.ui.preference.dsl.OptionsFragment
@@ -22,7 +22,7 @@ class DisplayPreferencesScreen : OptionsFragment() {
 	private val preferencesId by lazy { requireArguments().getString(ARG_PREFERENCES_ID) }
 	private val allowViewSelection by lazy { requireArguments().getBoolean(ARG_ALLOW_VIEW_SELECTION) }
 
-	override val stores: Array<IPreferenceStore>
+	override val stores: Array<PreferenceStore>
 		get() = arrayOf(libraryPreferences)
 
 	override val screen by optionsScreen {
