+++ app/src/main/java/org/jellyfin/androidtv/ui/preference/screen/HomePreferencesScreen.kt
@@ -2,7 +2,7 @@ package org.jellyfin.androidtv.ui.preference.screen
 
 import org.jellyfin.androidtv.R
 import org.jellyfin.androidtv.constant.HomeSectionType
-import org.jellyfin.androidtv.preference.IPreferenceStore
+import org.jellyfin.androidtv.preference.PreferenceStore
 import org.jellyfin.androidtv.preference.UserSettingPreferences
 import org.jellyfin.androidtv.ui.preference.dsl.OptionsFragment
 import org.jellyfin.androidtv.ui.preference.dsl.enum
@@ -12,7 +12,7 @@ import org.koin.android.ext.android.inject
 class HomePreferencesScreen : OptionsFragment() {
 	private val userSettingPreferences: UserSettingPreferences by inject()
 
-	override val stores: Array<IPreferenceStore>
+	override val stores: Array<PreferenceStore>
 		get() = arrayOf(userSettingPreferences)
 
 	override val screen by optionsScreen {
